Cleanup and simplify use of RPC reference counts
Previously, reference counts were only acquired when RPC locks
were released temporarily. This resulted in lots of code. Now,
references are taken by top-level functions that identify RPCs
to manipulate. Any function that receives an RPC as an argument
can assume that a reference has been taken on the RPC, so there
is no need to take a reference when an RPC lock is released.